tests/requirements.txt
grpcio~=1.73.0
grpcio-tools~=1.73.0
pytest~=8.4.0
pytest-asyncio

pip~=25.1.1
protobuf~=6.31.1
Pygments~=2.19.1
pluggy~=1.6.0
iniconfig~=2.1.0
packaging~=25.0
setuptools~=80.9.0
docker-py~=1.10.6
